𝕏 AI/requirements.txt
langchain>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0 